extract dialog into own component
